Structure analytics into distinct files
